fix: handle namespaced mutation correctly
